Removed separate throttle.ts
