+++ src/core-utils.ts
@@ -51,6 +51,22 @@ export function adjustmentsEqual(item1: SizeUnit[], item2: SizeUnit[]) {
 	return true;
 }
 
+export function throttle<F extends (...args: any) => any>(callback: F, limit: number) {
+	var wait = false; // Initially, we're not waiting
+	return function(...args: any) {
+		// We return a throttled function
+		if (!wait) {
+			// If we're not waiting
+			callback(...args); // Execute users function
+			wait = true; // Prevent future invocations
+			setTimeout(function() {
+				// After a period of time
+				wait = false; // And allow future invocations
+			}, limit);
+		}
+	};
+}
+
 export function styleDefinition(space: ISpaceDefinition) {
 	const style: React.CSSProperties = {
 		position: space.position,
